% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{blockRandom}
\alias{blockRandom}
\title{Derive a randomization of a table.}
\usage{
blockRandom(S, x = NA, check = TRUE)
}
\arguments{
\item{S}{a \code{data.frame}}

\item{x}{the column name of the group used for the block randomization.}

\item{check}{check if the size of each group is equal.}
}
\value{
returns a randomization order.
}
\description{
Derive a randomization of a table.
}
\examples{

set.seed(1)
iris[c(1:2, 51:52, 101:103), ] |>
  blockRandom(x = "Species", check=FALSE)
}
