\name{pgImporter}

\alias{pgImporter}

\title{Progenesis Importer Function}

\description{
    The function reads a CSV file which is exported using ProgenesisLCMS. The object is flexible when it comes to exported fields, ideally, everything is in the CSV. 
    The delimiter is a semicolon (this setting is potentially ProgenesisLCMS version or system language setting sensitive). All the information is kept and accessible.
}

\usage{

    pgImporter(file)

}

\arguments{
  \item{file}{A csv file.}
}

\details{

    t.b.d.
}

\author{

   Hubert Rehauer, Jonas Grossmann 2012

}
