\name{summary.mascot}
\alias{summary.mascot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Mascot Search Result Summary
}
\description{
  provides some properties of a mascot search result.
}
\usage{
  \method{summary}{mascot}(object, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
a mascot search result object \code{mascot}
  }
  \item{\dots}{
    pass parameters.
  }
}
\value{
  prints a summary.
}

\author{
  Christian Panse, 2017
}

\seealso{
\itemize{
  \item{\code{\link{plot.mascot_query}}}
}
}
